locked in and made the shader for the first level to use as a template
I honestly have no idea how i managed this. also all of the references in the controller script have to be placeholders until it can get meshed with the input system and ui and changed to trigger on events instead of on start. Hardest part i'll see with that is getting the coordinates for the reference pixel.